Use indirect filehandles and 3-argument "open" in {save,load}_filter
Use indirect filehandles (open $fh, ...) instead of simple glob
filehandle (open F, ...) in save_filter() and load_filter() to be able
to have them local to subroutines.

Use 3-argument version of "open" in save_filter() and load_filter() to
be safe against filenames containing characters which might influence
filemode etc.; it is a bit safer, and it is one string interpolation
less.

Signed-off-by: Jakub Narebski <jnareb@gmail.com>