Print filtered to $filter_window via 'Channel Message' text event
Print filtered content to specified window with nickname, via 'Channel
Message' text event, using Xchar::emit_print().  This allows for
exactly the same custom formatting as would be used in original window
(channel) if content was not filtered.

Just in case we print filtered content only if $nick is defined; this
might change in the future.

Signed-off-by: Jakub Narebski <jnareb@gmail.com>