New config: Add '/FILTER CONVERT [<file>]', to load old format config
Currently the same as '/FILTER (RE)LOAD', only with optional parameter
(which defaults to "$xchatdir/SoftSnow_filter.conf").  It is meant to
load old style list of DENY rules config file, to be able then to save
it in new config format (to be done).

While at it load_filter() subroutine now checks if the file exists at
all, before trying to open it.

Signed-off-by: Jakub Narebski <jnareb@gmail.com>